Fix potential bug regarding the use of env-variable
Prior to this change it was possible to create a conflict within AWS
since the codebuild-project was just named after the repository.  But
since we provide a "env"-variable is is effectively possible to setup
the very same project in one environment twice.  This resulted in a
conflict within AWS because we tried to use the same name twice as well.
This change fixes this problem.